requirements.txt
aiogram>=3.0
aiohttp
python-dotenv
apscheduler
beautifulsoup4
lxml
